sankofawatch/requirements.txt
# Web framework
Django>=4.2.0,<5.0
djangorestframework>=3.14.0
whitenoise>=6.5.0
django-cors-headers>=4.2.0
django-csp>=3.7
gunicorn>=21.0.0

# Database
psycopg2-binary>=2.9.0

# Config
python-decouple>=3.8
django-environ>=0.11.0

# Task queue
celery>=5.3.0
redis>=5.0.0
django-celery-beat>=2.5.0

# Geospatial
rasterio>=1.3.0
geopandas>=0.14.0
pyproj>=3.6.0
shapely>=2.0.0

# Earth Engine
earthengine-api>=0.1.0

# ML
torch>=2.0.0
torchvision>=0.15.0
segmentation-models-pytorch>=0.3.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
opencv-python-headless>=4.8.0

# Utilities
Pillow>=10.0.0
requests>=2.31.0
osmium>=3.7.0
